App/Part: add API Property::isSameContent()
A generate implementation of comparing property of any type. Used in
Part::PropertyGeometryList::isSame().

Note, in the future, isSame() will be promoted to be an abstract API in
Property, and every type of property will have to provide an
implementation. isSameContent() can be used as a not so efficient
fallback.