Core: fix import of FreeCAD in a Python 3.12 interpreter
Importing FreeCAD as a module in a normal Python interpreter (e.g. /usr/bin/python3) stopped working with Python 3.12 (but works normally with Python 3.11). When executing "import FreeCAD as App", the interpreter crashes with "Fatal Python error: PyImport_AppendInittab: PyImport_AppendInittab() may not be called after Py_Initialize()".
The check that causes the error message and abortion was introduced in Python 3.12 via
https://github.com/python/cpython/commit/7f3a4b967cfb1596a3fda6c34f900f8586b16700
(and a typo in the error message fixed via)
https://github.com/python/cpython/commit/35dd55005ee9aea2843eff7f514ee689a0995df8

Fixes #1073