App/Part: do not report error/warning if not recomputing
The fix in this commit is specifically for SubShapeBinder but is also
applicable to other objects.

SubShapeBinder may update its shape immediately when its bound shape
changes. It will update again when it is recomputed. The first update
may cause transient error that may be resolved before the second
recompute for various reason, e.g. when there is expression involved in
binder properties.